+++ apps/api/src/services/exchange-rate-data/exchange-rate-data.service.ts
@@ -30,6 +30,7 @@ import ms from 'ms';
 export class ExchangeRateDataService {
   private currencies: string[] = [];
   private currencyPairs: DataGatheringItem[] = [];
+  private derivedCurrencyFactors: { [currencyPair: string]: number } = {};
   private exchangeRates: { [currencyPair: string]: number } = {};
 
   public constructor(
@@ -135,7 +136,13 @@ export class ExchangeRateDataService {
   public async initialize() {
     this.currencies = await this.prepareCurrencies();
     this.currencyPairs = [];
+    this.derivedCurrencyFactors = {};
     this.exchangeRates = {};
+
+    for (const { currency, factor, rootCurrency } of DERIVED_CURRENCIES) {
+      this.derivedCurrencyFactors[`${currency}${rootCurrency}`] = 1 / factor;
+      this.derivedCurrencyFactors[`${rootCurrency}${currency}`] = factor;
+    }
 
     for (const {
       currency1,
@@ -266,10 +273,14 @@ export class ExchangeRateDataService {
       return this.toCurrency(aValue, aFromCurrency, aToCurrency);
     }
 
+    const derivedCurrencyFactor =
+      this.derivedCurrencyFactors[`${aFromCurrency}${aToCurrency}`];
     let factor: number;
 
     if (aFromCurrency === aToCurrency) {
       factor = 1;
+    } else if (derivedCurrencyFactor) {
+      factor = derivedCurrencyFactor;
     } else {
       const dataSource =
         this.dataProviderService.getDataSourceForExchangeRates();
@@ -357,7 +368,22 @@ export class ExchangeRateDataService {
       for (const date of dates) {
         factors[format(date, DATE_FORMAT)] = 1;
       }
-    } else {
+
+      return factors;
+    }
+
+    const derivedCurrencyFactor =
+      this.derivedCurrencyFactors[`${currencyFrom}${currencyTo}`];
+
+    if (derivedCurrencyFactor) {
+      for (const date of dates) {
+        factors[format(date, DATE_FORMAT)] = derivedCurrencyFactor;
+      }
+
+      return factors;
+    }
+
+    {
       const dataSource =
         this.dataProviderService.getDataSourceForExchangeRates();
       const symbol = `${currencyFrom}${currencyTo}`;