+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed an issue with the edit of future activities (drafts)
+- Fixed an issue with the exchange rate calculation when converting between derived currencies and their root currencies
 
 ## 2.218.0 - 2025-11-20